+++ lib/entry/v2/EntryUpdateCL.ts
@@ -127,7 +127,7 @@ export default class EntryUpdateCL<EntityT extends object = object, EntityKeysT 
                 content.setBindingContext(this.getEntryContext(), this.getModelName());
             }
 
-            fragment.open();
+            fragment.open(true);
         } else {
             fragment.destroyFragmentContent();
             throw new Error("Provided fragment must contain a sap.m.Dialog control. Put all the controls into a sap.m.Dialog");