+++ lib/ui/ContentCL.ts
@@ -320,6 +320,18 @@ export default class ContentCL<EntryT extends EntryCL<EntityT>, EntityT extends 
             smartField.setEditable(false);
         }
 
+        const customData = this.entry.getFieldCustomData().find(data => data.propertyName === property.propertyName);
+
+        if (customData) {
+            smartField.addCustomData(customData.customData);
+        }
+
+        const editMode = this.entry.getTextInEditModeSource().find(mode => mode.propertyName === property.propertyName);
+
+        if (editMode) {
+            smartField.setTextInEditModeSource(editMode.textInEditModeSource);
+        }
+
         const groupElement = new GroupElement({
             elements: [smartField]
         });
@@ -372,10 +384,16 @@ export default class ContentCL<EntryT extends EntryCL<EntityT>, EntityT extends 
 
         if (property.nullable === "false") {
             checkbox.setRequired(true);
-        }
+        }        
 
         if (this.entry.getReadonlyProperties().includes(property.propertyName)) {
             checkbox.setEditable(false);
+        }
+
+        const customData = this.entry.getFieldCustomData().find(data => data.propertyName === property.propertyName);
+
+        if (customData) {
+            checkbox.addCustomData(customData.customData);
         }
 
         this.simpleFormElements.push(checkbox);
@@ -410,6 +428,12 @@ export default class ContentCL<EntryT extends EntryCL<EntityT>, EntityT extends 
             datePicker.setEditable(false);
         }
 
+        const customData = this.entry.getFieldCustomData().find(data => data.propertyName === property.propertyName);
+
+        if (customData) {
+            datePicker.addCustomData(customData.customData);
+        }
+
         this.simpleFormElements.push(datePicker);
     }
 
@@ -436,6 +460,12 @@ export default class ContentCL<EntryT extends EntryCL<EntityT>, EntityT extends 
 
         if (this.entry.getReadonlyProperties().includes(property.propertyName)) {
             dateTimePicker.setEditable(false);
+        }
+
+        const customData = this.entry.getFieldCustomData().find(data => data.propertyName === property.propertyName);
+
+        if (customData) {
+            dateTimePicker.addCustomData(customData.customData);
         }
 
         this.simpleFormElements.push(dateTimePicker);
@@ -531,6 +561,12 @@ export default class ContentCL<EntryT extends EntryCL<EntityT>, EntityT extends 
             input.attachValueHelpRequest({}, valueHelp.openValueHelpDialog, valueHelp);
         }
 
+        const customData = this.entry.getFieldCustomData().find(data => data.propertyName === property.propertyName);
+
+        if (customData) {
+            input.addCustomData(customData.customData);
+        }
+
         this.simpleFormElements.push(input);
     }
 
@@ -539,7 +575,9 @@ export default class ContentCL<EntryT extends EntryCL<EntityT>, EntityT extends 
             elements: [control.getControl()]
         });
 
-        if (!this.entry.getUseMetadataLabels()) {
+        if (this.entry.getUseMetadataLabels()) {
+            groupElement.setLabel(property.annotationLabel || this.getEntityTypePropLabel(property.propertyName));
+        } else {
             groupElement.setLabel(this.getEntityTypePropLabel(property.propertyName));
         }